Revert "Add option to manually configure HVAC modes"
