+++ custom_components/heatmiserneo/switch.py
@@ -104,7 +104,8 @@ class NeoTimerEntity(CoordinatorEntity, SwitchEntity):
             "identifiers": {("heatmiser neoStat", self._timer.name)},
             "name": self._timer.name,
             "manufacturer": "Heatmiser",
-            "model": "neoTimer"
+            "model": "neoTimer",
+            "suggested_area": self._timer.name,
         }
     
     async def async_turn_on(self, **kwargs):
